fix(path-security): close NFKC fold-introduced separator gap in validateWithinRoots
validateWithinRoots only matched a fixed homoglyph character class, so a
character that NFKC-normalizes into a path separator but is outside that list
(U+2100 -> "a/c", U+FE68 -> "\") was accepted — even though sanitizeComponent
already rejected the same input as UNICODE_TRICK. A downstream layer that
normalizes the string (URL decode + NFKC, a name-folding filesystem) would then
see a separator the lexical guard never did.

Extract the NFKC detector into a shared unicode-trick module so the component
sanitizer and the path validator cannot drift. The detector is now count-based
(folding must not increase the separator count or introduce a "..") rather than
presence-based: a presence check is useless for a full path, which already
contains separators — that is exactly why the path validator missed these.

Adds two failing-before/passing-after regression vectors to the allowed-roots
suite (U+2100 forward slash, U+FE68 backslash).
